+++ build.gradle
@@ -28,7 +28,7 @@ minecraft {
     runDir = "run"
     mappings = "snapshot_20171003"
     def args = [
-            '-Dfml.coreMods.load=zone.rong.mixinbooter.MixinBooterPlugin',
+            '-Dfml.coreMods.load=zone.rong.mixinbooter.internal.MixinBooterPlugin',
             '-Dmixin.hotSwap=true',
             '-Dmixin.checks.interfaces=true'
     ]
@@ -108,10 +108,10 @@ publishing {
     // select the repositories you want to publish to
     repositories {
         maven {
-            url cleanroom_maven
+//            url cleanroom_maven
             credentials {
-                username cleanroom_user
-                password cleanroom_pass
+ //               username cleanroom_user
+  //              password cleanroom_pass
             }
         }
     }
