--- src/main/java/zone/rong/mixinbooter/api/IMixinLogGenerator.java
@@ -1,13 +1,0 @@
-package zone.rong.mixinbooter.api;
-
-import java.util.List;
-
-public interface IMixinLogGenerator {
-
-    List<String> getMixinConfigs();
-
-    boolean shouldMixinReportCustomMessage(String mixinConfig);
-
-    String onMixinMessage(String mixinConfig);
-
-}
