--- src/main/java/zone/rong/mixinbooter/api/MixinMessage.java
@@ -1,12 +1,0 @@
-package zone.rong.mixinbooter.api;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface MixinMessage {
-    String value();
-}
