+++ src/main/java/zone/rong/mixinbooter/internal/mixin/CrashReportMixin.java
@@ -1,4 +1,4 @@
-package zone.rong.mixinbooter.mixin;
+package zone.rong.mixinbooter.internal.mixin;
 
 import net.minecraft.crash.CrashReport;
 import org.spongepowered.asm.mixin.Mixin;
@@ -6,7 +6,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
-import zone.rong.mixinbooter.api.MixinStack;
+import zone.rong.mixinbooter.IMixinStack;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -16,7 +16,7 @@ public class CrashReportMixin {
 
     @Inject(method = "getCauseStackTraceOrString", at = @At("RETURN"), locals = LocalCapture.CAPTURE_FAILHARD, cancellable = true)
     private void afterStackTracePopulation(CallbackInfoReturnable<String> cir, StringWriter stringwriter, PrintWriter printwriter, Throwable throwable) {
-        cir.setReturnValue(cir.getReturnValue() + MixinStack.createStackReport(throwable).getStackMessage());
+        cir.setReturnValue(cir.getReturnValue() + IMixinStack.createStackReport(throwable).getStackMessage());
     }
 
 }
