+++ src/main/java/zone/rong/mixinbooter/internal/mixin/MixinFurnaceTile.java
@@ -1,4 +1,4 @@
-package zone.rong.mixinbooter.mixin;
+package zone.rong.mixinbooter.internal.mixin;
 
 import net.minecraft.tileentity.TileEntityFurnace;
 import org.spongepowered.asm.mixin.Mixin;
@@ -6,9 +6,10 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import zone.rong.mixinbooter.api.MixinMessage;
+import zone.rong.mixinbooter.annotations.MixinMessage;
 
 @Mixin(TileEntityFurnace.class)
+@MixinMessage("Class mixin message")
 public class MixinFurnaceTile {
 
     @Inject(method = "update",at = @At("HEAD"))
