+++ src/main/java/zone/rong/mixinbooter/internal/mixin/MixinWorldServer.java
@@ -1,11 +1,11 @@
-package zone.rong.mixinbooter.mixin;
+package zone.rong.mixinbooter.internal.mixin;
 
 import net.minecraft.world.WorldServer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import zone.rong.mixinbooter.api.MixinMessage;
+import zone.rong.mixinbooter.annotations.MixinMessage;
 
 @Mixin(WorldServer.class)
 public class MixinWorldServer {
